requirements.txt
SQLAlchemy~=1.4.44
fastapi~=0.75.0
pydantic~=1.9.0
pymysql~=1.0.2
uvicorn~=0.19.0
python-jose[cryptography]
passlib[bcrypt]~=1.7.4
python-multipart~=0.0.5
pytz~=2021.3